refactor: change event capacity type from BigInteger to Integer

+++ src/main/java/com/github/rafael_souza_de_almeida/ticket_mania/catalog/domain/Event.java
@@ -31,6 +31,6 @@ public class Event {
     private String place;
 
     @Column(nullable = false)
-    private BigInteger capacity;
+    private Integer capacity;
 
 }
