+++ src/main/java/com/github/rafael_souza_de_almeida/ticket_mania/catalog/dto/EventResponseDto.java
@@ -12,7 +12,7 @@ public class EventResponseDto {
     private String name;
     private LocalDateTime date;
     private String place;
-    private BigInteger capacity;
+    private Integer capacity;
 
     public EventResponseDto(Event event) {
         this.name = event.getName();
